Added "Emergency Mode" to process item display.
If the show global checkbox is checked, then double-clicking on the
globe icon shows all of the process items of that type.

This can be useful in case the user is looking for an item which has
a process scope, but can't remember the process name.

This is not documented right now. It would be easy to update the globe
icon title when show global is checked to make it (slightly) discoverable.